Fixes internal transitions for parent states
The correction for issue17 resulted in internal transitions in parent
states not being executed when nested states are involved.
When an event is not handled by a child state but can be accepted by an
internal transition in a parent state, the actions associated with the
internal state should be executed but without leaving the child state.